+++ changes.md
@@ -16,6 +16,7 @@
 - `Twig.lib.date` -> `require('locutus/php/datetime/date')`
 - `Twig.lib.boolval` -> `require('locutus/php/var/boolval')`
 - `Twig.Template` -> `require('./model/Template')`
+- `Twig.prepare` -> `require('./helper/prepare')`
 # Remove
 - `Twig.lib`